Show pipeline total execution time in a new status dot at the bottom
