+++ js_modules/dagit/src/execution/RunMetadataProvider.tsx
@@ -28,6 +28,7 @@ export interface IRunMetadataDict {
   startingProcessAt?: number;
   startedProcessAt?: number;
   startedPipelineAt?: number;
+  exitedAt?: number;
   processId?: number;
   steps: {
     [stepName: string]: IStepMetadata;
@@ -51,6 +52,12 @@ function extractMetadataFromLogs(
     }
     if (log.__typename === "PipelineStartEvent") {
       metadata.startedPipelineAt = Number.parseInt(log.timestamp);
+    }
+    if (
+      log.__typename === "PipelineFailureEvent" ||
+      log.__typename === "PipelineSuccessEvent"
+    ) {
+      metadata.exitedAt = Number.parseInt(log.timestamp);
     }
 
     if ("step" in log) {
